Add content-twitter/store/tweet.ts
