+++ src/content-twitter/component/scrapbox-button.tsx
@@ -42,7 +42,7 @@ export const ScrapboxButton: React.FC<ScrapboxButtonProps> = ({ tweetID }) => {
   const tweetRef = React.useRef(null);
   // redux
   const selector = React.useCallback(
-    (state: State) => state[toTweetIDKey(tweetID)],
+    (state: State) => state.tweet[toTweetIDKey(tweetID)],
     [tweetID],
   );
   const buttonState = useSelector(selector);
@@ -60,14 +60,16 @@ export const ScrapboxButton: React.FC<ScrapboxButtonProps> = ({ tweetID }) => {
   // click: add tweet
   const onClickAddTweet = async () => {
     // change to in-progress
-    dispatch(actions.update({ tweetID, state: { state: 'in-progress' } }));
+    dispatch(
+      actions.tweet.update({ tweetID, state: { state: 'in-progress' } }),
+    );
     // add tweet
     const result = await addTweet(tweetID, tweetRef.current, logger);
     if (result.ok) {
       setTooltipMessage({ type: 'notification', message: 'Copied' });
     } else {
       dispatch(
-        actions.update({
+        actions.tweet.update({
           tweetID,
           state: { state: 'failure', message: result.error },
         }),
@@ -76,7 +78,7 @@ export const ScrapboxButton: React.FC<ScrapboxButtonProps> = ({ tweetID }) => {
         type: 'error',
         message: `Failed: ${result.error}`,
         onClosed: () =>
-          dispatch(actions.update({ tweetID, state: { state: 'none' } })),
+          dispatch(actions.tweet.update({ tweetID, state: { state: 'none' } })),
       });
     }
     // show result with tooltip
@@ -85,7 +87,9 @@ export const ScrapboxButton: React.FC<ScrapboxButtonProps> = ({ tweetID }) => {
   // click: delete tweet
   const onClickDeleteTweet = async () => {
     // change to in-progress
-    dispatch(actions.update({ tweetID, state: { state: 'in-progress' } }));
+    dispatch(
+      actions.tweet.update({ tweetID, state: { state: 'in-progress' } }),
+    );
     // delete tweet
     const result = await deleteTweet(tweetID, logger);
     if (result.ok) {
@@ -95,7 +99,9 @@ export const ScrapboxButton: React.FC<ScrapboxButtonProps> = ({ tweetID }) => {
         type: 'error',
         message: `Failed: ${result.error}`,
         onClosed: () =>
-          dispatch(actions.update({ tweetID, state: { state: 'success' } })),
+          dispatch(
+            actions.tweet.update({ tweetID, state: { state: 'success' } }),
+          ),
       });
     }
     // show result with tooltip
