+++ src/content-twitter/index.tsx
@@ -12,7 +12,8 @@ import { savedTweetIDs } from '~/lib/storage/tweet';
 import { ScrapboxButton } from './component/scrapbox-button';
 import './index.scss';
 import { insertReactRoot } from './lib/insert-react-root';
-import { UpdateButtonState, actions, store } from './store';
+import { actions, store } from './store';
+import { UpdateButtonState } from './store/tweet';
 
 logger.info('content script');
 
@@ -28,7 +29,7 @@ const observerCallback = (records: MutationRecord[]): void => {
       insertReactRoot(node, new URL(document.URL), logger).forEach(
         ({ tweetID, reactRoot }) => {
           // update store
-          store.dispatch(actions.touch(tweetID));
+          store.dispatch(actions.tweet.touch(tweetID));
           // render by React
           const root = createRoot(reactRoot);
           root.render(
@@ -57,7 +58,7 @@ window.addEventListener('DOMContentLoaded', () => {
   savedTweetIDs().then((tweetIDs) => {
     logger.debug('Saved Tweet IDs', tweetIDs);
     store.dispatch(
-      actions.update(
+      actions.tweet.update(
         tweetIDs.map((tweetID) => ({ tweetID, state: { state: 'success' } })),
       ),
     );
@@ -92,7 +93,7 @@ const storageListener = createStorageListener(
     }
     // update state
     if (buttonStates.length) {
-      store.dispatch(actions.update(buttonStates));
+      store.dispatch(actions.tweet.update(buttonStates));
     }
   },
   logger,
